+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+ - Added overloads to RPC server for calling a method on all connected clients
+
 ### Changed
  - MockMqttClient now "publishes" messages to an internal class. Added console command for viewing messages.